fix: do not ser return_url to allow auto-watch
This does not really fix the issue (as described in #9). It’s not possible to navigate to about:blank page through browserstack API - we would have to use WebDriver instead, which is in general significantly slower.

I tested this a couple of times and it looks like the issue does not exist with BS (we used to have this problem with SL) - BS really kills the browser before the terminateWorker response is sent so I hope the re-capturing old browsers should not be an issue.

Closes #9