fix(launcher) workers fail to launch after tunnel
In some environments, the process which establishes the tunnel can
report that it is operational significantly before the tunnel is
actually established. This will cause subsequent calls to create a
worker to fail.

This change waits a short interval after the tunnel process has
reported itself started to allow it to actually finish establishing
the tunnel before attempting to create any workers.

Fixes karma-runner/karma-browserstack-launcher#42