Permit overriding `name`, `build`, and `project` on a per-browser basis.
Browserstack uses the `name` property in the browser configuration
to uniquely name sessions within a specific build. If these sessions
do not have a unique name, and are executed sequentially (or with a
concurrency < total # of browsers), browserstack's UI will overwrite
previous results with new ones. In cases where a karma test run is
executed on multiple browsers, this isn't really desirable.

While I was at it, I also incorporated changes from
https://github.com/karma-runner/karma-browserstack-launcher/pull/71, and
updated the README to also include documentation for that change. With
this patch, `name`, `project`, and `build`, may all be overridden on
a per browser basis.